Added new sourcer68 skeleton.

+++ mksc68/mksc68_gdb.c
@@ -1179,9 +1179,9 @@ gdb_t * gdb_create(void)
     /* $$$ TEMP: options need to be set with gdb command
        ATM breaks on all timers but C (system timer)
     */
-    gdb->opt.ontimer_break = ~0x44;
+    gdb->opt.ontimer_break = (uint8_t)~0x44;
     /* $$$ only once */
-    gdb->opt.ontimer_break = ~0xf4;
+    gdb->opt.ontimer_break = (uint8_t)~0xf4;
   }
   return gdb;
